+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
-## Upcoming
+## v2026.27 - 2026-04-22
+* Security: Fix personal note sharing permissions (https://github.com/Syslifters/sysreptor/security/advisories/GHSA-pcpr-q2qj-3v43)
 * Image editor: add pixelation tool
 * MDE: support toggling task list checkboxes in markdown preview
 * MDE: support frontmatter highlighting